rename thread instances, data directory and ticker name

+++ src/llmq/quorums_dkgsessionhandler.cpp
@@ -145,7 +145,7 @@ void CDKGSessionHandler::StartThread()
         throw std::runtime_error("Tried to start an already started CDKGSessionHandler thread.");
     }
 
-    std::string threadName = strprintf("q-phase-%d", params.type);
+    std::string threadName = strprintf("q-phase-%d", (int8_t)params.type);
     phaseHandlerThread = std::thread(&TraceThread<std::function<void()> >, threadName, std::function<void()>(std::bind(&CDKGSessionHandler::PhaseHandlerThread, this)));
 }
